feat: home skeleton
