+++ src/core/components/page-wrapper/BlitzClientWrapper.tsx
@@ -1,0 +1,13 @@
+// app/BlitzClientWrapper.tsx
+
+'use client'
+
+import { BlitzProvider } from '../../../app/blitz-client'
+
+export default function BlitzClientWrapper({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return <BlitzProvider>{children}</BlitzProvider>
+}
