+++ src/core/layouts/Layout.tsx
@@ -1,3 +1,5 @@
+// TODO: DELETE THIS FILE
+
 'use client'
 
 import React from 'react'
